requirements.txt
playwright==1.44.0
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
